backend/admin-backend/pyproject.toml
[project]
name = "admin-backend"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<4.1.0",
    "python-jose[cryptography]>=3.3.0",
    "sqlmodel>=0.0.21",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "redis>=5.2.0",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
    "greenlet>=3.0.0",
    "apscheduler>=3.10.4",
]
